requirements.txt
alpaca_trade_api
bs4
pandas
pendulum
python-dotenv
requests
scikit-learn
tensorflow
yfinance
